+++ ExilePartyClient/src/app/authorize/components/char-profile/char-profile.module.ts
@@ -4,13 +4,21 @@ import { CharInventoryModule } from './char-inventory/char-inventory.module';
 import { CharEquipmentModule } from './char-equipment/char-equipment.module';
 import { SharedModule } from '../../../shared/shared.module';
 import { CharSummaryModule } from './char-summary/char-summary.module';
+import { MatTabsModule } from '@angular/material/tabs';
+import { CharWealthModule } from './char-wealth/char-wealth.module';
+import { MatIconModule } from '@angular/material/icon';
+import { CharMapsModule } from './char-maps/char-maps.module';
 
 @NgModule({
   imports: [
     SharedModule,
     CharInventoryModule,
     CharEquipmentModule,
-    CharSummaryModule
+    CharSummaryModule,
+    CharWealthModule,
+    CharMapsModule,
+    MatTabsModule,
+    MatIconModule
   ],
   declarations: [CharProfileComponent],
   exports: [CharProfileComponent]