+++ ExilePartyClient/src/app/authorize/components/char-profile/char-wealth/char-wealth.component.ts
@@ -1,0 +1,15 @@
+import { Component, OnInit } from '@angular/core';
+
+@Component({
+  selector: 'app-char-wealth',
+  templateUrl: './char-wealth.component.html',
+  styleUrls: ['./char-wealth.component.scss']
+})
+export class CharWealthComponent implements OnInit {
+
+  constructor() { }
+
+  ngOnInit() {
+  }
+
+}