+++ ExilePartyClient/src/app/authorize/inspect-players/inspect-players.module.ts
@@ -1,7 +1,7 @@
 import { NgModule } from '@angular/core';
 
 import { SharedModule } from '../../shared/shared.module';
-import { PartyComponent } from './party.component';
+import { InspectPlayersComponent } from './inspect-players.component';
 import { PlayerListModule } from './../components/player-list/player-list.module';
 import { CharProfileModule } from '../components/char-profile/char-profile.module';
 import {MatDividerModule} from '@angular/material/divider';
@@ -13,7 +13,7 @@ import { LogMonitorService } from '../../shared/providers/log-monitor.service';
     PlayerListModule,
     CharProfileModule
   ],
-  declarations: [PartyComponent],
+  declarations: [InspectPlayersComponent],
   providers: [LogMonitorService]
 })
-export class PartyModule { }
+export class InspectPlayersModule { }